More fixes

+++ sign_stream.go
@@ -170,7 +170,7 @@ func (s *signAttachedStream) signBlock(isFinal bool) error {
 		return err
 	}
 
-	assertEncodedChunkState(s.version, chunk, 0, s.seqno, isFinal)
+	assertEncodedChunkState(s.version, chunk, 0, uint64(s.seqno), isFinal)
 
 	sBlock := makeSignatureBlock(s.version, sig, chunk, isFinal)
 	if err := s.encoder.Encode(sBlock); err != nil {
